refactor(code-preview): move curated syntax manifest to shared preview code module
