+++ src/preview/code/backends/mod.rs
@@ -1,4 +1,3 @@
 pub(super) mod custom;
 pub(super) mod plain;
 pub(super) mod syntect;
-pub(super) mod syntect_manifest;
